// client/src/app/records.ts
export class Records {
    _id?: String;
    patient_id: String;
    date: String;
    nurse_name: String;
    type: String;
    category: String;
    reading1: String;
    reading2: String;
    food_allergy: String;
    bp: String;
    diabetic: String;
    heart_disease: String;
    surgery: String;
    accident: String;
}